Delete search router and controller
