--- back/src/routers/search.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { getSearchedProductTitles } from "../controllers/search";
-
-const searchRoutes:Router = Router();
-
-searchRoutes.get('/:searchQuery',  getSearchedProductTitles)
-
-export default searchRoutes;
