+++ front/src/axios.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { TypeProduct } from "./types";
 import { formatQuery } from "./utils";
-import { CATEGORY_URL, PRODUCT_URL, SEARCH_URL } from "./configs/axios_urls";
+import { CATEGORY_URL, PRODUCT_URL, PRODUCT_SEARCH_URL } from "./configs/axios_urls";
 
 export const getCategoriesNames = async (): Promise<string[]> => {
     try {
@@ -47,7 +47,7 @@ export const getSearchedProductNames = async (query: string, setResults: React.D
     setLoading(true);
     try {
         const format_query = formatQuery(query);
-        const response = await axios.get(`${SEARCH_URL}/${format_query}`);
+        const response = await axios.get(`${PRODUCT_SEARCH_URL}${format_query}`);
         const axios_result: TypeProduct[] = response.data.products;
         setResults(axios_result.length ? axios_result : [{ title: "Not found" }]);
     } catch (error) {
