+++ front/src/configs/axios_urls.ts
@@ -2,4 +2,4 @@ export const CATEGORY_URL = `${process.env.REACT_APP_AXIOS_HOST}/category`
 
 export const PRODUCT_URL = `${process.env.REACT_APP_AXIOS_HOST}/product`
 
-export const SEARCH_URL = `${process.env.REACT_APP_AXIOS_HOST}/search`
+export const PRODUCT_SEARCH_URL = `${process.env.REACT_APP_AXIOS_HOST}/product?search=`
